requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
line-bot-sdk>=3.5.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
aiosqlite>=0.19.0
pytz>=2023.3
httpx>=0.25.0
